pyproject.toml
[project]
name = "py-ffmpeg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Urrutia Mickael", email = "mickael.urrutia@utc.fr" }
]
requires-python = ">=3.13"
dependencies = [
    "py-utils",
    "python-ffmpeg>=2.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyside6-essentials>=6.9.1",
    "python-dateutil>=2.9.0.post0",
]

[tool.uv.sources]
py-utils = { path = "../py-utils", editable = true }
